chore(core): Remove remnants of conflict resolution from Document
